+++ kotti_image/tests/conftest.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 Created on 2015-12-16
 :author: Andreas Kaiser (disko@binary-punks.com)
@@ -10,15 +8,18 @@ pytest_plugins = "kotti"
 from pytest import fixture
 
 
-@fixture(scope='session')
+@fixture(scope="session")
 def custom_settings():
-    # Compatibility for Kotti 1.3 and 2.0
-    try:
-        from kotti.resources import Image
-        return {}
-    except ImportError:
-        # noinspection PyUnresolvedReferences c
-        from kotti_image.resources import Image
-        return {
-            'kotti.configurators': 'kotti_image.kotti_configure'
-        }
+    # noinspection PyUnresolvedReferences
+    from kotti_image.resources import Image
+
+    return {
+        "kotti.configurators": "kotti_image.kotti_configure",
+    }
+
+
+@fixture
+def app(db_session, setup_app):
+    from webtest import TestApp
+
+    return TestApp(setup_app)