+++ kotti_image/views/__init__.py
@@ -1,18 +1,15 @@
-# -*- coding: utf-8 -*-
-
 """
 Created on 2015-12-16
 :author: Andreas Kaiser (disko@binary-punks.com)
 """
 
 import PIL
-import rfc6266_parser
+from depot.utils import make_content_disposition
 from kotti.util import extract_from_settings
 from plone.scale.scale import scaleImage
 from pyramid.response import Response
 from pyramid.view import view_config
 from pyramid.view import view_defaults
-from unidecode import unidecode
 
 from kotti_image.interfaces import IImage
 
@@ -20,22 +17,22 @@ PIL.ImageFile.MAXBLOCK = 33554432
 
 #: Default image scales
 image_scales = {
-    'span1': [60, 120],
-    'span2': [160, 320],
-    'span3': [260, 520],
-    'span4': [360, 720],
-    'span5': [460, 920],
-    'span6': [560, 1120],
-    'span7': [660, 1320],
-    'span8': [760, 1520],
-    'span9': [860, 1720],
-    'span10': [960, 1920],
-    'span11': [1060, 2120],
-    'span12': [1160, 2320],
-    }
+    "span1": [60, 120],
+    "span2": [160, 320],
+    "span3": [260, 520],
+    "span4": [360, 720],
+    "span5": [460, 920],
+    "span6": [560, 1120],
+    "span7": [660, 1320],
+    "span8": [760, 1520],
+    "span9": [860, 1720],
+    "span10": [960, 1920],
+    "span11": [1060, 2120],
+    "span12": [1160, 2320],
+}
 
 
-@view_defaults(context=IImage, permission='view')
+@view_defaults(context=IImage, permission="view")
 class ImageView(object):
     """The ImageView class is registered for the :class:`IImage` context."""
 
@@ -44,7 +41,7 @@ class ImageView(object):
         self.context = context
         self.request = request
 
-    @view_config(name='view', renderer='kotti_image:templates/image.pt')
+    @view_config(name="view", renderer="kotti_image:templates/image.pt")
     def view(self):
         """
         :result: Empty dictionary to be handed to the image.pt template for
@@ -54,7 +51,7 @@ class ImageView(object):
 
         return {}
 
-    @view_config(name='image')
+    @view_config(name="image")
     def image(self, subpath=None):
         """Return the image in a specific scale, either inline
         (default) or as attachment.
@@ -79,11 +76,11 @@ class ImageView(object):
         width, height = (None, None)
         subpath = list(subpath)
 
-        if (len(subpath) > 0) and (subpath[-1] == 'download'):
-            disposition = 'attachment'
+        if (len(subpath) > 0) and (subpath[-1] == "download"):
+            disposition = "attachment"
             subpath.pop()
         else:
-            disposition = 'inline'
+            disposition = "inline"
 
         if len(subpath) == 1:
             scale = subpath[0]
@@ -92,35 +89,39 @@ class ImageView(object):
                 width, height = image_scales[scale]
 
         if not (width and height):
-            return self.request.uploaded_file_response(
-                self.context.data, disposition)
+            return self.request.uploaded_file_response(self.context.data, disposition)
 
-        image, format, size = scaleImage(self.context.data.file.read(),
-                                         width=width,
-                                         height=height,
-                                         direction='thumb')
+        image, format, size = scaleImage(
+            self.context.data.file.read(), width=width, height=height, mode="contain"
+        )
         res = Response(
             headerlist=[
-                ('Content-Disposition', '{0};filename="{1}"'.format(
-                    disposition,
-                    self.context.filename.encode('ascii', 'ignore'))),
-                ('Content-Length', str(len(image))),
-                ('Content-Type', str(self.context.mimetype)),
+                (
+                    "Content-Disposition",
+                    '{0};filename="{1}"'.format(
+                        disposition, self.context.filename.encode("ascii", "ignore")
+                    ),
+                ),
+                ("Content-Length", str(len(image))),
+                ("Content-Type", str(self.context.mimetype)),
             ],
             body=image,
         )
-        res.content_disposition = rfc6266_parser.build_header(
-            self.context.filename, disposition=disposition,
-            filename_compat=unidecode(self.context.filename))
+        # res.content_disposition = rfc6266_parser.build_header(
+        #     self.context.filename, disposition=disposition,
+        #     filename_compat=unidecode(self.context.filename))
+        res.content_disposition = make_content_disposition(
+            disposition, self.context.filename
+        )
 
         return res
 
 
 def _load_image_scales(settings):
-    image_scale_strings = extract_from_settings('kotti.image_scales.', settings)
+    image_scale_strings = extract_from_settings("kotti.image_scales.", settings)
 
     for k in image_scale_strings.keys():
-        image_scales[k] = [int(x) for x in image_scale_strings[k].split('x')]
+        image_scales[k] = [int(x) for x in image_scale_strings[k].split("x")]
 
 
 def includeme(config):