+++ kotti_image/views/edit.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 Created on 2015-12-16
 :author: Andreas Kaiser (disko@binary-punks.com)
@@ -12,15 +10,21 @@ from kotti_image import _
 from kotti_image.resources import Image
 
 
-@view_config(name=Image.type_info.add_view,
-             permission=Image.type_info.add_permission,
-             renderer='kotti:templates/edit/node.pt')
+@view_config(
+    name=Image.type_info.add_view,
+    permission=Image.type_info.add_permission,
+    renderer="kotti:templates/edit/node.pt",
+)
 class ImageAddForm(FileAddForm):
-    item_type = _(u"Image")
+    item_type = _("Image")
     item_class = Image
 
 
-@view_config(name='edit', context=Image, permission='edit',
-             renderer='kotti:templates/edit/node.pt')
+@view_config(
+    name="edit",
+    context=Image,
+    permission="edit",
+    renderer="kotti:templates/edit/node.pt",
+)
 class ImageEditForm(FileEditForm):
     pass